test: . move SimulatedFile and related functions to a dedicated file
